+++ Pooling.h
@@ -9,10 +9,10 @@ class PoolingLayer : public Layer {
 
         PoolingLayer(int input_chan, int output_chan, PoolingType type, int pool_size, int stride, Mode mode = REGULAR); 
 
-        Tensor<double, 4> forward(const Tensor<double, 4> &input) override;
+        Tensor<float, 4> forward(const Tensor<float, 4> &input) override;
 
         void backward(const MatrixXd& grad) override {}
-        void update(double learnign_rate) override {}
+        void update(float learnign_rate) override {}
     private:
         PoolingType type;
         int pool_size;
